rewrite torch save
