+++ neural_compressor/compression/__init__.py
@@ -16,5 +16,5 @@
 # limitations under the License.
 
 from .callbacks import QuantizationAwareTrainingCallbacks, DistillationCallbacks, PruningCallbacks
-from .pruner.pruning import PruningWrapper, PruningUnWrapper
+from ..experimental.compression import prepare_pruning
 from .. import WeightPruningConfig
