+++ neural_compressor/experimental/compression/__init__.py
@@ -1,0 +1,1 @@
+from .pruning import prepare_pruning
